Error handling

+++ src/controllers/pasteController.js
@@ -6,10 +6,14 @@ const ObjectID = require('mongodb').ObjectId;
 
 
 //@desc  create Paste
-//@route POST /Pastes
+//@route POST /Pastes 
 //@access public
 const createPaste = asyncHandler( async (req, res) => {
     const { title, code, language, password } = req.body;
+    if(!title || !code || !language){
+        res.status(400);
+        throw new Error("All fields are mandatory !");
+    }
     let link
     while(1){
         link = makelink(10)
@@ -18,7 +22,7 @@ const createPaste = asyncHandler( async (req, res) => {
             break;
     }
     const paste = await Paste.create({
-        link: link,
+        link: link, 
         title,
         code,
         language,
@@ -35,18 +39,23 @@ const getPaste = asyncHandler( async (req, res) => {
     const paste = await Paste.find({ link: req.params.id}).setOptions({ sanitizeFilter: true });
     if(paste.length == 0){
         // error
-        res.status(404).json("Not found");
+        res.status(404);
         throw new Error("Paste not Found");
     }
     const hashedpass = paste[0]["password"]
-    bcrypt.compare(password, hashedpass, function(error, match){
-        if(match){
-            res.status(200).json(paste)
-        }
-        else{
-            res.status(403).json("Access Denied")
-        }
-    })
+    if(!hashedpass){
+        res.status(200).json(paste)
+    }
+    else{
+        bcrypt.compare(password, hashedpass, function(error, match){
+            if(!match){
+                res.status(403).json('Unauthorized')
+            }
+            else{
+                res.status(200).json(paste)
+            }
+        })
+    }
 });
 
 //@desc  Update Paste
@@ -57,33 +66,38 @@ const updatePaste = asyncHandler( async (req, res) => {
     const paste = await Paste.find({ link: req.params.id}).setOptions({ sanitizeFilter: true });
     if(paste.length == 0){
         // error
-        res.status(404).json("Not found");
-        throw new Error("Paste not Found");
+        res.status(404)
+        throw new Error("Paste not Found")
     }
     const hashedpass = paste[0]["password"] 
-    bcrypt.compare(password, hashedpass, async function(error, match){
-        if(error){
-            throw error
-        }
-        else if(!match){
-            res.status(403).json("Access Denied")
-        }
-        else{
-            const ID=paste[0]._id;
-            const updatedPaste= await Paste.findByIdAndUpdate(
-                ID,
-                {
-                    link: req.params.id,
-                    title: title,
-                    code: code,
-                    language: language,
-                    password: hashedpass
-                },
-                {new: true}
-            );
-            res.status(200).json(updatedPaste);
-        }
-    })
+    if(!hashedpass){
+        res.status(403).json('Unauthorized')
+    }
+    else{
+        bcrypt.compare(password, hashedpass, async function(error, match){
+            if(error){
+                throw error
+            }
+            else if(!match){
+                res.status(403).json('Unauthorized')
+            }
+            else{
+                const ID=paste[0]._id;
+                const updatedPaste= await Paste.findByIdAndUpdate(
+                    ID,
+                    {
+                        link: req.params.id,
+                        title: title,
+                        code: code,
+                        language: language,
+                        password: hashedpass
+                    },
+                    {new: true}
+                );
+                res.status(200).json(updatedPaste);
+            }
+        })
+    }
 });
 
 //@desc  delete Paste
@@ -94,27 +108,31 @@ const deletePaste = asyncHandler( async (req, res) => {
     const paste = await Paste.find({ link: req.params.id}).setOptions({ sanitizeFilter: true });
     if(paste.length == 0){
         // error
-        res.status(404).json("Not found");
-        throw new Error("Paste not Found");
+        res.status(404)
+        throw new Error("Paste not Found")
     }
     const hashedpass = paste[0]["password"]
-    bcrypt.compare(password, hashedpass, async function(error, match){
-        if(error){
-            throw error  
-        }
-        else if(!match){
-            res.status(403).json("Access Denied")
-        }
-        else{
-            const ID = paste[0]._id 
-            Paste.deleteOne({_id:new  
-                ObjectID(ID)}).then(data=>{
-                console.log("deleted")
-            })
-            res.status(200).json("Deleted")
-        }
-    })
-    
+    if(!hashedpass){
+        res.status(403).json('Unauthorized')
+    }
+    else{
+        bcrypt.compare(password, hashedpass, async function(error, match){
+            if(error){
+                throw error  
+            }
+            else if(!match){
+                res.status(403).json('Unauthorized')
+            }
+            else{
+                const ID = paste[0]._id 
+                Paste.deleteOne({_id:new  
+                    ObjectID(ID)}).then(data=>{
+                    console.log("deleted")
+                })
+                res.status(200).json("Deleted")
+            }
+        })
+    }
 });
 
 module.exports = {createPaste, getPaste, updatePaste, deletePaste}
